Bouton sup added + popup bug regler + modifier fixed

+++ CalendarEnMieux/model/eventsManagement.php
@@ -23,6 +23,19 @@ function showData($date){
     return $result;
 }
 
+function showAllData($userID){
+  $result = false;
+
+  $alldateAgenda = "(SELECT EVENTS.date FROM EVENTS WHERE FKusers = '$userID') UNION
+  (SELECT `event-recurrence`.date FROM `event-recurrence` INNER JOIN events ON EVENTS.ID = `event-recurrence`.FKevents  WHERE FKusers = '$userID')";
+
+  require_once 'model/dbConnector.php';
+
+  $result = executeQuerySelect($alldateAgenda);
+
+  return $result;
+}
+
 function selectAllEvents($date){
 
 
@@ -35,7 +48,7 @@ function selectAllEvents($date){
 }
 function selectAllRecurrence($date){
 
-    $selectAllRecurenceQuery="SELECT * FROM (SELECT ID, name, place, `start time`, `end time`,`type`,recurrence FROM events ) AS events
+    $selectAllRecurenceQuery="SELECT events.ID, name, place, `start time`, `end time`,`type`,recurrence, `event-recurrence`.ID AS recuID ,`event-recurrence`.date, FKevents FROM (SELECT ID, name, place, `start time`, `end time`,`type`,recurrence FROM events ) AS events
     right JOIN `event-recurrence` ON events.ID = `event-recurrence`.FKevents WHERE date = '$date'";
 
    	require_once 'model/dbConnector.php';
@@ -56,6 +69,7 @@ function addEvent($eventToAdd, $userID){
     $endTime = $eventToAdd['endTime'];
     $type = $eventToAdd['type'];
     $recurrence = $eventToAdd['recurrence'];
+    $qty = $eventToAdd['qty'];
 
     $i=0;
     if(isset($recurrence[$i])){
@@ -81,7 +95,7 @@ function addEvent($eventToAdd, $userID){
 
       if($recurrence == 1){
         //jours
-        for($i=1;$i<=60;$i++){
+        for($qty;$qty>=0;$qty--){
 
           $date = new DateTime("{$date}");
           $date->add(new DateInterval('P1D'));
@@ -91,12 +105,12 @@ function addEvent($eventToAdd, $userID){
           $addEventQuery2 = 'INSERT INTO `event-recurrence` (`date`, `FKevents`) VALUES (:date, :FKevents)';
           $addEventData2 = array(":date" => $date, ":FKevents" => $idEvent);
 
-          $result = executeQueryInsert($addEventQuery2,$addEventData2);
+          $result2 = executeQueryInsert($addEventQuery2,$addEventData2);
         }
       }
       //mois
       if($recurrence == 2){
-        for($i=1;$i<=30;$i++){
+        for($qty;$qty>=0;$qty--){
           $date = new DateTime("{$date}");
           $date->add(new DateInterval('P1M'));
           $date -> format('Y-m-d');
@@ -105,12 +119,12 @@ function addEvent($eventToAdd, $userID){
           $addEventQuery2 = 'INSERT INTO `event-recurrence` (`date`, `FKevents`) VALUES (:date, :FKevents)';
           $addEventData2 = array(":date" => $date, ":FKevents" => $idEvent);
 
-          $result = executeQueryInsert($addEventQuery2,$addEventData2);
+          $result2 = executeQueryInsert($addEventQuery2,$addEventData2);
         }
       }
       //année
       if($recurrence == 3){
-        for($i=1;$i<=10;$i++){
+        for($qty;$qty>=0;$qty--){
           $date = new DateTime("{$date}");
           $date->add(new DateInterval('P1Y'));
           $date -> format('Y-m-d');
@@ -119,7 +133,7 @@ function addEvent($eventToAdd, $userID){
           $addEventQuery2 = 'INSERT INTO `event-recurrence` (`date`, `FKevents`) VALUES (:date, :FKevents)';
           $addEventData2 = array(":date" => $date, ":FKevents" => $idEvent);
 
-          $result = executeQueryInsert($addEventQuery2,$addEventData2);
+          $result2 = executeQueryInsert($addEventQuery2,$addEventData2);
         }
       }
 
@@ -129,11 +143,55 @@ function addEvent($eventToAdd, $userID){
 
 function deleteEvent($eventToDelete, $userID){
 
-    $suppQuery='DELETE from events where ID = :id AND FKusers = :idUser';
-    $suppData= array(":id" => $eventToDelete['sup'], ":idUser" => $userID);
+    $date = $eventToDelete["date"];
 
     require_once 'model/dbConnector.php';
-    $result = executeQueryInsert($suppQuery, $suppData);
+
+    //supprimer uniquement l'event choisi
+    if($eventToDelete["sup"] != ""){
+
+      if($eventToDelete["recurrence"] != 0){
+
+        $suppQuery2='DELETE from `event-recurrence` where ID = :id';
+        $suppData2= array(":id" => $eventToDelete['sup']);
+
+        print_r($suppQuery2);
+        print_r($suppData2);
+
+        $result = executeQueryInsert($suppQuery2, $suppData2);
+
+      }
+      else{
+        $suppQuery='DELETE from events where ID = :id AND FKusers = :idUser';
+        $suppData= array(":id" => $eventToDelete['sup'], ":idUser" => $userID);
+
+        $result = executeQueryInsert($suppQuery, $suppData);
+      }
+
+    }
+    //supprimer l'event et toutes les recurrences
+    if($eventToDelete["supAll"] != ""){
+
+      $suppQuery2='DELETE from `event-recurrence` where FKevents = :id';
+      $suppData2= array(":id" => $eventToDelete['supAll']);
+
+      $result2 = executeQueryInsert($suppQuery2, $suppData2);
+
+      $suppQuery='DELETE from events where ID = :id AND FKusers = :idUser';
+      $suppData= array(":id" => $eventToDelete['supAll'], ":idUser" => $userID);
+
+      $result = executeQueryInsert($suppQuery, $suppData);
+
+    }
+    //supprimer l'event choisi et les suivantes recurrences
+    if($eventToDelete["supAfter"] != ""){
+
+      $suppQuery="DELETE FROM `event-recurrence` WHERE FKevents = :id AND date > '$date' ";
+      $suppData= array(":id" => $eventToDelete['supAfter']);
+
+      $result = executeQueryInsert($suppQuery, $suppData);
+
+    }
 
     return $result;
 }
@@ -148,6 +206,7 @@ function updateEvent($eventToModify,$userID){
     $type = $eventToModify['type'];
     $recurrence = $eventToModify['recurrence'];
     $IDOfEvent = $eventToModify['upd'];
+    $qty = $eventToModify['qty'];
 
     $strSeparator = '\'';
 
@@ -163,9 +222,68 @@ function updateEvent($eventToModify,$userID){
     WHERE ID = '.$strSeparator.$IDOfEvent.$strSeparator.' AND FKusers = '.$strSeparator.$userID.$strSeparator;
     $updateEventData = array(":name" => $event, ":place" => $place, ":date" => $date, ":startTime" => $startTime, ":endTime" => $endTime, ":type" => $type, ":recurrence" => $recurrence, ":userID" => $userID);
 
-
     require_once 'model/dbConnector.php';
     $result = executeQueryInsert($updateEventQuery, $updateEventData);
+
+    $idEvent = $IDOfEvent;
+
+    $selectAllEventsQuery="SELECT recurrence FROM events WHERE ID ='$idEvent'";
+   	require_once 'model/dbConnector.php';
+
+   	$oldRecurrence = executeQuerySelect($selectAllEventsQuery);
+
+
+    if($oldRecurrence != $recurrence){
+
+      $suppOldRecurrenceQuery='DELETE from `event-recurrence` WHERE FKevents = :id';
+      $suppOldRecurrenceData= array(":id" => $idEvent);
+
+      $result3 = executeQueryInsert($suppOldRecurrenceQuery, $suppOldRecurrenceData);
+
+      if($recurrence == 1){
+        //jours
+        for($qty;$qty>=1;$qty--){
+
+          $date = new DateTime("{$date}");
+          $date->add(new DateInterval('P1D'));
+          $date -> format('Y-m-d');
+          $date = $date->format('Y-m-d');
+
+          $addEventQuery2 = 'INSERT INTO `event-recurrence` (`date`, `FKevents`) VALUES (:date, :FKevents)';
+          $addEventData2 = array(":date" => $date, ":FKevents" => $idEvent);
+
+          $result2 = executeQueryInsert($addEventQuery2,$addEventData2);
+        }
+      }
+      //mois
+      if($recurrence == 2){
+        for($qty;$qty>=1;$qty--){
+          $date = new DateTime("{$date}");
+          $date->add(new DateInterval('P1M'));
+          $date -> format('Y-m-d');
+          $date = $date->format('Y-m-d');
+
+          $addEventQuery2 = 'INSERT INTO `event-recurrence` (`date`, `FKevents`) VALUES (:date, :FKevents)';
+          $addEventData2 = array(":date" => $date, ":FKevents" => $idEvent);
+
+          $result2 = executeQueryInsert($addEventQuery2,$addEventData2);
+        }
+      }
+      //année
+      if($recurrence == 3){
+        for($qty;$qty>=1;$qty--){
+          $date = new DateTime("{$date}");
+          $date->add(new DateInterval('P1Y'));
+          $date -> format('Y-m-d');
+          $date = $date->format('Y-m-d');
+
+          $addEventQuery2 = 'INSERT INTO `event-recurrence` (`date`, `FKevents`) VALUES (:date, :FKevents)';
+          $addEventData2 = array(":date" => $date, ":FKevents" => $idEvent);
+
+          $result2 = executeQueryInsert($addEventQuery2,$addEventData2);
+        }
+      }
+    }
 
     return $result;
 }
